# core/planifai-core/src/main/resources/openapi/core-openapi.yml
openapi: 3.0.3
info:
  title: PlanifAI Core API
  version: 0.1.0
  description: Contract-first API for PlanifAI core modules.
servers:
  - url: http://localhost:8082
paths:
...
  /api/finance/expenses:
    get:
      tags:
        - Finance
      operationId: getExpenses
      summary: List expenses
      parameters:
        - name: category
          in: query
          required: false
          description: Optional backend-controlled finance category filter.
          schema:
            $ref: '#/components/schemas/FinanceCategory'
      responses:
        '200':
          description: Expenses returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpenseResponse'
    post:
      tags:
        - Finance
      operationId: createExpense
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseRequest'
      responses:
        '201':
          description: Expense created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
  /api/finance/transactions:
    get:
      tags:
        - Finance
      operationId: getFinanceTransactions
      summary: List finance transactions
      description: Lists transaction-like finance records currently represented by expenses. This endpoint is read-only and supports category filtering.
      parameters:
        - name: category
          in: query
          required: false
          description: Optional backend-controlled finance category filter.
          schema:
            $ref: '#/components/schemas/FinanceCategory'
      responses:
        '200':
          description: Finance transactions returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpenseResponse'
  /api/finance/incomes:
    get:
      tags:
        - Finance
      operationId: getIncomes
      summary: List incomes
      responses:
        '200':
          description: Incomes returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IncomeResponse'
    post:
      tags:
        - Finance
      operationId: createIncome
      summary: Create income
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncomeRequest'
      responses:
        '201':
          description: Income created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomeResponse'
  /api/finance/dashboard:
    get:
      tags:
        - Finance
      operationId: getFinanceDashboard
      summary: Get monthly finance dashboard
      parameters:
        - name: month
          in: query
          required: true
          description: Month to summarize using YYYY-MM format.
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
            example: '2026-05'
      responses:
        '200':
          description: Monthly finance dashboard returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceDashboardResponse'
...
  /api/finance/statistics/categories:
    get:
      tags:
        - Finance
      operationId: getFinanceCategoryStatistics
      summary: Get expense statistics by category
      description: Returns deterministic monthly expense totals grouped by backend-controlled finance category. Categories without expense movement are omitted.
      parameters:
        - name: month
          in: query
          required: true
          description: Month to summarize using YYYY-MM format.
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
            example: '2026-05'
      responses:
        '200':
          description: Category statistics returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceCategoryStatisticsResponse'
  /api/finance/budgets:
    get:
      tags:
        - Finance
      operationId: getFinanceBudgets
      summary: List monthly category budgets
      description: Lists manually managed category budgets for a month. Budgets are planning records only and do not synchronize bank accounts, trigger payments, send push notifications, or perform AI categorization.
      parameters:
        - name: month
          in: query
          required: true
          description: Month to list using YYYY-MM format.
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
            example: '2026-05'
      responses:
        '200':
          description: Monthly budgets returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BudgetResponse'
    post:
      tags:
        - Finance
      operationId: createFinanceBudget
      summary: Create monthly category budget
      description: Creates a manually tracked monthly budget for a backend-controlled finance category. This endpoint does not create banking rules, push notifications, or automatic categorization.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetRequest'
      responses:
        '201':
          description: Budget created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetResponse'
  /api/finance/budgets/summary:
    get:
      tags:
        - Finance
      operationId: getFinanceBudgetSummary
      summary: Get monthly budget summary
      description: Returns monthly budget consumption grouped by category. Summary values are computed from manually tracked budgets and finance expenses only; no bank synchronization, push alerts, or AI categorization are triggered.
      parameters:
        - name: month
          in: query
          required: true
          description: Month to summarize using YYYY-MM format.
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
            example: '2026-05'
      responses:
        '200':
          description: Monthly budget summary returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetSummaryResponse'
  /api/finance/budgets/{id}:
    get:
      tags:
        - Finance
      operationId: getFinanceBudgetById
      summary: Get budget by id
      description: Returns a manually tracked monthly category budget with computed consumption fields.
      parameters:
        - name: id
          in: path
          required: true
          description: Budget identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Budget returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetResponse'
        '404':
          description: Budget not found.
    put:
      tags:
        - Finance
      operationId: updateFinanceBudget
      summary: Update budget
      description: Updates a manually tracked monthly category budget. This does not affect bank accounts, notifications, or AI categorization.
      parameters:
        - name: id
          in: path
          required: true
          description: Budget identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetRequest'
      responses:
        '200':
          description: Budget updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetResponse'
        '404':
          description: Budget not found.
    delete:
      tags:
        - Finance
      operationId: deleteFinanceBudget
      summary: Delete budget
      description: Deletes a manually tracked monthly category budget. This does not delete expenses or change banking data.
      parameters:
        - name: id
          in: path
          required: true
          description: Budget identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '204':
          description: Budget deleted.
        '404':
          description: Budget not found.
  /api/finance/savings-goals:
    get:
      tags:
        - Finance
      operationId: getSavingsGoals
      summary: List savings goals
      description: Lists medium and long-term savings goals tracked manually in PlanifAI. This endpoint does not create investment products, simulate interest, synchronize bank accounts, or manage shared goals.
      responses:
        '200':
          description: Savings goals returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavingsGoalResponse'
    post:
      tags:
        - Finance
      operationId: createSavingsGoal
      summary: Create savings goal
      description: Creates a manually tracked savings goal. The goal is informational only and does not trigger investments, interest simulations, shared goal flows, or banking integrations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavingsGoalRequest'
      responses:
        '201':
          description: Savings goal created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoalResponse'
...
  /api/finance/savings-goals/{id}:
    get:
      tags:
        - Finance
      operationId: getSavingsGoalById
      summary: Get savings goal by id
      description: Returns a manually tracked savings goal with progress and ETA fields.
      parameters:
        - name: id
          in: path
          required: true
          description: Savings goal identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Savings goal returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoalResponse'
        '404':
          description: Savings goal not found.
    put:
      tags:
        - Finance
      operationId: updateSavingsGoal
      summary: Update savings goal
      description: Updates a manually tracked savings goal. This does not trigger investments, interest simulations, shared goal flows, or banking integrations.
      parameters:
        - name: id
          in: path
          required: true
          description: Savings goal identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavingsGoalRequest'
      responses:
        '200':
          description: Savings goal updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoalResponse'
        '404':
          description: Savings goal not found.
    delete:
      tags:
        - Finance
      operationId: deleteSavingsGoal
      summary: Delete savings goal
      description: Deletes a manually tracked savings goal from PlanifAI. This does not affect bank accounts, investments, or shared goals.
      parameters:
        - name: id
          in: path
          required: true
          description: Savings goal identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '204':
          description: Savings goal deleted.
        '404':
          description: Savings goal not found.
  /api/finance/recurring-expenses:
    get:
      tags:
        - Finance
      operationId: getRecurringExpenses
      summary: List recurring expenses
      description: Lists manually managed recurring expenses and financial obligations. This endpoint does not create automatic payments and does not synchronize with banks.
      parameters:
        - name: category
          in: query
          required: false
          description: Optional backend-controlled finance category filter.
          schema:
            $ref: '#/components/schemas/FinanceCategory'
      responses:
        '200':
          description: Recurring expenses returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecurringExpenseResponse'
    post:
      tags:
        - Finance
      operationId: createRecurringExpense
      summary: Create recurring expense
      description: Creates a manually tracked recurring expense. No payment is executed and no bank synchronization is triggered.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringExpenseRequest'
      responses:
        '201':
          description: Recurring expense created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringExpenseResponse'
  /api/finance/recurring-expenses/{id}:
    put:
      tags:
        - Finance
      operationId: updateRecurringExpense
      summary: Update recurring expense
      description: Updates a manually tracked recurring expense. No payment is executed and no bank synchronization is triggered.
      parameters:
        - name: id
          in: path
          required: true
          description: Recurring expense identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringExpenseRequest'
      responses:
        '200':
          description: Recurring expense updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringExpenseResponse'
        '404':
          description: Recurring expense not found.
    delete:
      tags:
        - Finance
      operationId: deleteRecurringExpense
      summary: Delete recurring expense
      description: Deletes a manually tracked recurring expense from PlanifAI. This does not cancel external subscriptions, execute payments, or synchronize with banks.
      parameters:
        - name: id
          in: path
          required: true
          description: Recurring expense identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '204':
          description: Recurring expense deleted.
        '404':
          description: Recurring expense not found.
  /api/finance/obligations/monthly-summary:
    get:
      tags:
        - Finance
      operationId: getMonthlyObligationsSummary
      summary: Get monthly obligations summary
      description: Projects manually tracked recurring obligations for a month. The projection is informational only; PlanifAI does not execute payments and does not synchronize with bank accounts.
      parameters:
        - name: month
          in: query
          required: true
          description: Month to project using YYYY-MM format.
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
            example: '2026-05'
      responses:
        '200':
          description: Monthly obligations summary returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlyObligationsSummaryResponse'
components:
  schemas:
...
    FinanceCategory:
      type: string
      description: Backend-controlled finance category used for expenses, transaction-like records, recurring expenses, projections, and dashboards. Dynamic user-created categories are intentionally deferred.
      enum: [FOOD, RESTAURANTS, TRANSPORT, HEALTH, EDUCATION, SUBSCRIPTIONS, SAVINGS, ENTERTAINMENT, TRAVEL, PETS, OTHER, HOUSING, UTILITIES]
...
    FinanceCategoryStatisticsResponse:
      type: object
      required: [month, totalExpenses, categories]
      properties:
        month:
          type: string
          pattern: '^\d{4}-(0[1-9]|1[0-2])$'
          description: Month summarized using YYYY-MM format.
          example: '2026-05'
        totalExpenses:
          type: number
          format: double
          minimum: 0
          description: Total expenses for the selected month.
        categories:
          type: array
          description: Expense categories with movement in the selected month. Empty months return an empty array.
          items:
            $ref: '#/components/schemas/FinanceCategoryStatisticItem'
...
    Recurrence:
      type: string
      enum: [ONE_OFF, MONTHLY, YEARLY]
...
    FinanceDashboardResponse:
      type: object
      required:
        - month
        - totalIncome
        - totalExpenses
        - netBalance
        - savingsAmount
        - savingsRate
        - healthStatus
        - expensesByCategory
      properties:
        month:
          type: string
          pattern: '^\d{4}-(0[1-9]|1[0-2])$'
          description: Month summarized by the dashboard using YYYY-MM format.
          example: '2026-05'
        totalIncome:
          type: number
          format: double
          description: Total income for the selected month.
          minimum: 0
        totalExpenses:
          type: number
          format: double
          description: Total expenses for the selected month.
          minimum: 0
        netBalance:
          type: number
          format: double
          description: Total income minus total expenses.
        savingsAmount:
          type: number
          format: double
          description: Amount saved in the selected month.
        savingsRate:
          type: number
          format: double
          description: Savings percentage over total income. Empty-income months return 0 and healthStatus NO_DATA. Negative-balance months may return a negative savings rate.
          maximum: 100
        healthStatus:
          $ref: '#/components/schemas/FinancialHealthStatus'
        expensesByCategory:
          type: array
          description: Expense totals grouped by category. Empty months return an empty array.
          items:
            $ref: '#/components/schemas/ExpenseCategoryBreakdown'
    BudgetRequest:
      type: object
      description: Manual monthly budget definition for one backend-controlled finance category. This DTO does not create bank rules, push notifications, or AI categorization instructions.
      required: [month, category, limitAmount, active]
      properties:
        month:
          type: string
          pattern: '^\d{4}-(0[1-9]|1[0-2])$'
          description: Budget month using YYYY-MM format.
          example: '2026-05'
        category:
          $ref: '#/components/schemas/FinanceCategory'
        limitAmount:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
          description: Maximum planned spend for the category in the selected month.
        active:
          type: boolean
          description: Whether this budget should be included in monthly summary calculations.
        notes:
          type: string
          maxLength: 1000
          description: Optional user notes for manual budgeting context.
...
    BudgetResponse:
      allOf:
        - $ref: '#/components/schemas/BudgetRequest'
        - type: object
          required:
            - id
            - consumedAmount
            - remainingAmount
            - overspentAmount
            - consumptionPercentage
            - status
            - alerts
          properties:
            id:
              type: integer
              format: int64
              description: Budget identifier.
            consumedAmount:
              type: number
              format: double
              minimum: 0
              description: Expenses consumed for this category in the budget month.
            remainingAmount:
              type: number
              format: double
              minimum: 0
              description: Remaining amount before reaching limitAmount. Exceeded budgets return 0.
            overspentAmount:
              type: number
              format: double
              minimum: 0
              description: Amount above limitAmount. Non-exceeded budgets return 0.
            consumptionPercentage:
              type: number
              format: double
              minimum: 0
              description: Percentage of limitAmount consumed by expenses in this category.
            status:
              $ref: '#/components/schemas/BudgetStatus'
            alerts:
              type: array
              description: Informational alerts for this budget. Empty when no warning or exceeded condition exists.
              items:
                $ref: '#/components/schemas/BudgetAlert'
...
    BudgetSummaryResponse:
      type: object
      description: Monthly budget summary across active category budgets. This summary is informational and does not send notifications, synchronize banks, or modify expenses.
      required:
        - month
        - totalLimitAmount
        - totalConsumedAmount
        - totalRemainingAmount
        - totalOverspentAmount
        - overallConsumptionPercentage
        - status
        - categories
        - alerts
      properties:
        month:
          type: string
          pattern: '^\d{4}-(0[1-9]|1[0-2])$'
          description: Month summarized using YYYY-MM format.
          example: '2026-05'
        totalLimitAmount:
          type: number
          format: double
          minimum: 0
          description: Sum of active budget limits for the selected month.
        totalConsumedAmount:
          type: number
          format: double
          minimum: 0
          description: Sum of consumed expenses across active budget categories.
        totalRemainingAmount:
          type: number
          format: double
          minimum: 0
          description: Sum of remaining amounts across active budgets.
        totalOverspentAmount:
          type: number
          format: double
          minimum: 0
          description: Sum of overspent amounts across exceeded budgets.
        overallConsumptionPercentage:
          type: number
          format: double
          minimum: 0
          description: Percentage of totalLimitAmount consumed across active budgets. Months without active budget limits return 0.
        status:
          $ref: '#/components/schemas/BudgetStatus'
        categories:
          type: array
          description: Category-level budget statuses. Months without active budgets return an empty array.
          items:
            $ref: '#/components/schemas/BudgetCategoryStatus'
        alerts:
          type: array
          description: Aggregated informational alerts across the monthly budget summary.
          items:
            $ref: '#/components/schemas/BudgetAlert'
...
    RecurringExpenseResponse:
      allOf:
        - $ref: '#/components/schemas/RecurringExpenseRequest'
        - type: object
          required: [id]
          properties:
            id:
              type: integer
              format: int64
              description: Recurring expense identifier.
...
    ExpenseRequest:
      type: object
      required: [concept, amount, expenseDate, category]
      properties:
        concept:
          type: string
          minLength: 1
          maxLength: 160
        amount:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
        expenseDate:
          type: string
          format: date
        category:
          $ref: '#/components/schemas/FinanceCategory'
        recurrence:
          $ref: '#/components/schemas/Recurrence'
        notes:
          type: string
          maxLength: 1000
    ExpenseResponse:
      allOf:
        - $ref: '#/components/schemas/ExpenseRequest'
        - type: object
          required: [id, category, recurrence]
          properties:
            id:
              type: integer
              format: int64
    IncomeRequest:
      type: object
      required: [source, amount, incomeDate]
      properties:
        source:
          type: string
          minLength: 1
          maxLength: 160
        amount:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
        incomeDate:
          type: string
          format: date
        category:
          $ref: '#/components/schemas/IncomeCategory'
        recurrence:
          $ref: '#/components/schemas/Recurrence'
        notes:
          type: string
          maxLength: 1000
    IncomeResponse:
      allOf:
        - $ref: '#/components/schemas/IncomeRequest'
        - type: object
          required: [id, category, recurrence]
          properties:
            id:
              type: integer
              format: int64